+++ src/game/world/World.h
@@ -18,22 +18,22 @@ struct World {
     uint8_t id;
     glm::vec3 spawn_point;
 
-    std::unordered_map<uint8_t, std::unique_ptr<Chunk> > chunks{};
+    std::unordered_map<uint32_t, std::unique_ptr<Chunk> > chunks{};
 
     World(uint8_t id, const glm::vec3 &spawn_point) : id(id), spawn_point(spawn_point) {
     }
 
-    static constexpr int chunk_index(const int x, const int y, const int z) {
+    static constexpr uint32_t chunk_index(const uint32_t x, const uint32_t y, const uint32_t z) {
         return x + (WORLD_SIZE_X * y) + (WORLD_SIZE_X * WORLD_SIZE_Y * z);
     }
 
     [[nodiscard]] Chunk &getChunk(int x, int y, int z);
 
-    [[nodiscard]] Chunk &getChunk(uint8_t chunk_index);
+    [[nodiscard]] Chunk &getChunk(uint32_t chunk_index);
 
-    [[nodiscard]] bool isChunkLoaded(uint8_t chunk_index);
+    [[nodiscard]] bool isChunkLoaded(uint32_t chunk_index);
 
-    bool loadChunk(uint8_t chunk_index);
+    bool loadChunk(uint32_t chunk_index);
 
     static constexpr bool isOutOfBounds(const int x, const int y, const int z) {
         return x >= WORLD_SIZE_X || y >= WORLD_SIZE_Y || z >= WORLD_SIZE_Z || x < 0 || y < 0 || z < 0;
